Fixes for tooltip
